# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Yone wi",
        "description": "Cet api permet de gerer des reseaux de transports en commun au senegal et le partage d'informations sur les les abonnements ,trajets et leurs tarifs  ",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/types": {
            "get": {
                "tags": [
                    "Gestion des types"
                ],
                "summary": "Lister les types",
                "description": "",
                "operationId": "5ae5a6ccb7da5e1410da3b56b6fde09f",
                "parameters": [
                    {
                        "name": "User-Agent",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Gestion des types"
                ],
                "summary": "Ajouter un type",
                "description": "",
                "operationId": "6215f40499a4b69db488634fcb4ed0b4",
                "parameters": [
                    {
                        "name": "User-Agent",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorize"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/mestypes": {
            "get": {
                "tags": [
                    "Gestion des types"
                ],
                "summary": "Lister mes types",
                "description": "",
                "operationId": "afb380f09e091dd98a1264adb06fe0a8",
                "parameters": [
                    {
                        "name": "User-Agent",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/types/{type}": {
            "get": {
                "tags": [
                    "Gestion des types"
                ],
                "summary": "Afficher un type",
                "description": "",
                "operationId": "8c05f2c12bb06945e3c321a10e35af99",
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "User-Agent",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Gestion des types"
                ],
                "summary": "Supprimer un type",
                "description": "",
                "operationId": "05cb32f1c9e4f79fc0ea0057c9634071",
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "User-Agent",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorize"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Gestion des types"
                ],
                "summary": "Modifier un type",
                "description": "",
                "operationId": "06dbf1252262fa9f6702008db6b97415",
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "User-Agent",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorize"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/types/delete/{type}": {
            "patch": {
                "tags": [
                    "Gestion des types"
                ],
                "summary": "supprimer un type de la corbeille",
                "description": "",
                "operationId": "a922caec7998b77243f0e79baece679d",
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "User-Agent",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/types/restaurer/{type}": {
            "patch": {
                "tags": [
                    "Gestion des types"
                ],
                "summary": "restaurer un type",
                "description": "",
                "operationId": "3a4d137a71aaf4dfca560e70d3f867fd",
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "User-Agent",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/types/deleted/all": {
            "get": {
                "tags": [
                    "Gestion des types"
                ],
                "summary": "Lister les types qui sont dans la corbeille",
                "description": "",
                "operationId": "248f9e2e6c144abcc1c7491e4f3de6fe",
                "parameters": [
                    {
                        "name": "User-Agent",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/types/empty-trash": {
            "post": {
                "tags": [
                    "Gestion des types"
                ],
                "summary": "vider les types qui sont dans la corbeille",
                "description": "",
                "operationId": "225e6a80ce2b241ee218b942ce7c0363",
                "parameters": [
                    {
                        "name": "User-Agent",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorize"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}